export to string test fixed; json type adapters moved to separate packages
