+++ src/main/java/hwr/oop/group4/todo/persistence/json/adapters/deserializers/DateTimeDeserializer.java
@@ -1,0 +1,20 @@
+package hwr.oop.group4.todo.persistence.json.adapters.deserializers;
+
+import com.google.gson.JsonDeserializationContext;
+import com.google.gson.JsonDeserializer;
+import com.google.gson.JsonElement;
+import com.google.gson.JsonParseException;
+
+import java.lang.reflect.Type;
+import java.time.Instant;
+import java.time.LocalDateTime;
+import java.time.ZoneOffset;
+
+public class DateTimeDeserializer implements JsonDeserializer<LocalDateTime> {
+
+    @Override
+    public LocalDateTime deserialize(JsonElement jsonElement, Type type, JsonDeserializationContext jsonDeserializationContext) throws JsonParseException {
+        Instant instant = Instant.ofEpochMilli(jsonElement.getAsLong());
+        return LocalDateTime.ofInstant(instant, ZoneOffset.UTC);
+    }
+}
