+++ src/main/java/hwr/oop/group4/todo/persistence/json/adapters/serializers/DateTimeSerializer.java
@@ -1,0 +1,18 @@
+package hwr.oop.group4.todo.persistence.json.adapters.serializers;
+
+import com.google.gson.JsonElement;
+import com.google.gson.JsonPrimitive;
+import com.google.gson.JsonSerializationContext;
+import com.google.gson.JsonSerializer;
+
+import java.lang.reflect.Type;
+import java.time.LocalDateTime;
+import java.time.ZoneOffset;
+
+public class DateTimeSerializer implements JsonSerializer<LocalDateTime> {
+
+    @Override
+    public JsonElement serialize(LocalDateTime src, Type typeOfSrc, JsonSerializationContext context) {
+        return new JsonPrimitive(src.toInstant(ZoneOffset.UTC).toEpochMilli());
+    }
+}
