+++ src/test/java/hwr/oop/group4/todo/persistence/PersistableTodoListTest.java
@@ -53,12 +53,10 @@ class PersistableTodoListTest {
 
         Persistable persistedTodoList = new PersistableTodoList(todoList);
 
-        System.out.println(persistedTodoList.exportAsString());
-
         assertThatJson(persistedTodoList.exportAsString())
                 .when(Option.IGNORING_ARRAY_ORDER)
                 // TODO: use todolist.json
-                .isEqualTo("{\"maybeList\":[],\"projects\":[{\"name\":\"project a\",\"description\":\"\",\"tasks\":[{\"name\":\"task c\",\"description\":\"description c\",\"priority\":0,\"tags\":[],\"status\":\"OPEN\"},{\"name\":\"task b\",\"description\":\"description b\",\"priority\":0,\"tags\":[],\"status\":\"OPEN\"}],\"tags\":[]}],\"inTray\":[{},{}],\"loseTasks\":[{\"name\":\"task a\",\"description\":\"description a\",\"priority\":0,\"tags\":[{},{}],\"status\":\"OPEN\"}]}");
+                .isEqualTo("{\"projects\":[{\"name\":\"project a\",\"description\":\"\",\"tags\":[],\"tasks\":[{\"name\":\"task c\",\"description\":\"description c\",\"priority\":0,\"tags\":[],\"status\":\"OPEN\"},{\"name\":\"task b\",\"description\":\"description b\",\"priority\":0,\"tags\":[],\"status\":\"OPEN\"}]}],\"inTray\":[{\"name\":\"idea a\",\"description\":\"description a\"},{\"name\":\"idea b\",\"description\":\"description b\"}],\"loseTasks\":[{\"name\":\"task a\",\"description\":\"description a\",\"priority\":0,\"deadline\":1063004940000,\"tags\":[{\"name\":\"tagA\"},{\"name\":\"tagB\"}],\"status\":\"OPEN\"}],\"somedayMaybe\":[]}");
     }
 
     @Test
